Changed Longs to BigIntegers and BigDecimals. Stored as BigDecimals in db instead of bytes or strings.
1 Long to Int question still unanswered.

*Need to check forging algorthm. Forging is too fast now, much faster than stated block time in Constants.java.

Removed unused code for functions like
ALIAS
ASSET (Coloured Coins)
DIGITAL GOODS
ACCOUNT CONTROL
PHASING
POLL

+++ src/java/nxt/Block.java
@@ -43,9 +43,9 @@ public interface Block {
 
     long getNextBlockId();
 
-    long getTotalAmountNQT();
+    BigInteger getTotalAmountNQT();
 
-    long getTotalFeeNQT();
+    BigInteger getTotalFeeNQT();
 
     int getPayloadLength();
 
@@ -57,7 +57,7 @@ public interface Block {
 
     byte[] getBlockSignature();
 
-    long getBaseTarget();
+    BigInteger getBaseTarget();
 
     BigInteger getCumulativeDifficulty();
     
@@ -69,7 +69,7 @@ public interface Block {
     
     BigInteger getVault();
     
-    long getBlockReward();
+    BigInteger getBlockReward();
 
     byte[] getBytes();
